+++ src/bin/main.rs
@@ -1,4 +1,4 @@
-use great_calculator::{data_parser::Tree, *};
+use great_calculator::{parser::Tree, *, our_math_lib::*};
